fix(android): guard recovered writeback mutations
